+++ clustercontroller-reindexer/src/main/java/ai/vespa/reindexing/Reindexer.java
@@ -217,12 +217,10 @@ public class Reindexer {
     static class Cluster {
 
         private final String name;
-        private final String configId;
         private final Map<DocumentType, String> documentBuckets;
 
-        Cluster(String name, String configId, Map<DocumentType, String> documentBuckets) {
+        Cluster(String name, Map<DocumentType, String> documentBuckets) {
             this.name = requireNonNull(name);
-            this.configId = requireNonNull(configId);
             this.documentBuckets = Map.copyOf(documentBuckets);
         }
 
@@ -231,7 +229,7 @@ public class Reindexer {
         }
 
         String route() {
-            return "[Storage:cluster=" + name + ";clusterconfigid=" + configId + "]";
+            return name + "-direct";
         }
 
         String bucketSpaceOf(DocumentType documentType) {
@@ -244,20 +242,18 @@ public class Reindexer {
             if (o == null || getClass() != o.getClass()) return false;
             Cluster cluster = (Cluster) o;
             return name.equals(cluster.name) &&
-                   configId.equals(cluster.configId) &&
                    documentBuckets.equals(cluster.documentBuckets);
         }
 
         @Override
         public int hashCode() {
-            return Objects.hash(name, configId, documentBuckets);
+            return Objects.hash(name, documentBuckets);
         }
 
         @Override
         public String toString() {
             return "Cluster{" +
                    "name='" + name + '\'' +
-                   ", configId='" + configId + '\'' +
                    ", documentBuckets=" + documentBuckets +
                    '}';
         }
