+++ documentapi/src/main/java/com/yahoo/documentapi/messagebus/protocol/RoutingPolicyFactories.java
@@ -16,15 +16,6 @@ public abstract class RoutingPolicyFactories {
         }
     }
 
-    static class StoragePolicyFactory implements RoutingPolicyFactory {
-        public DocumentProtocolRoutingPolicy createPolicy(String param) {
-            return new StoragePolicy(param);
-        }
-
-        public void destroy() {
-        }
-    }
-
     static class ContentPolicyFactory implements RoutingPolicyFactory {
         public DocumentProtocolRoutingPolicy createPolicy(String param) {
             return new ContentPolicy(param);
