+++ documentapi/src/vespa/documentapi/messagebus/policies/CMakeLists.txt
@@ -3,7 +3,6 @@ vespa_add_library(documentapi_documentapipolicies OBJECT
     SOURCES
     andpolicy.cpp
     externslobrokpolicy.cpp
-    storagepolicy.cpp
     contentpolicy.cpp
     messagetypepolicy.cpp
     documentrouteselectorpolicy.cpp
