+++ documentapi/src/vespa/documentapi/messagebus/routingpolicyfactories.cpp
@@ -1,16 +1,15 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 #include "routingpolicyfactories.h"
 #include <vespa/documentapi/messagebus/policies/andpolicy.h>
+#include <vespa/documentapi/messagebus/policies/contentpolicy.h>
 #include <vespa/documentapi/messagebus/policies/documentrouteselectorpolicy.h>
 #include <vespa/documentapi/messagebus/policies/errorpolicy.h>
 #include <vespa/documentapi/messagebus/policies/externpolicy.h>
+#include <vespa/documentapi/messagebus/policies/loadbalancerpolicy.h>
 #include <vespa/documentapi/messagebus/policies/localservicepolicy.h>
+#include <vespa/documentapi/messagebus/policies/messagetypepolicy.h>
 #include <vespa/documentapi/messagebus/policies/roundrobinpolicy.h>
 #include <vespa/documentapi/messagebus/policies/subsetservicepolicy.h>
-#include <vespa/documentapi/messagebus/policies/storagepolicy.h>
-#include <vespa/documentapi/messagebus/policies/contentpolicy.h>
-#include <vespa/documentapi/messagebus/policies/messagetypepolicy.h>
-#include <vespa/documentapi/messagebus/policies/loadbalancerpolicy.h>
 
 using namespace documentapi;
 
@@ -18,17 +17,6 @@ mbus::IRoutingPolicy::UP
 RoutingPolicyFactories::AndPolicyFactory::createPolicy(const string &param) const
 {
     return mbus::IRoutingPolicy::UP(new ANDPolicy(param));
-}
-
-mbus::IRoutingPolicy::UP
-RoutingPolicyFactories::StoragePolicyFactory::createPolicy(const string &param) const
-{
-    mbus::IRoutingPolicy::UP ret(new StoragePolicy(param));
-    string error = static_cast<StoragePolicy&>(*ret).getError();
-    if (!error.empty()) {
-        ret.reset(new ErrorPolicy(error));
-    }
-    return ret;
 }
 
 mbus::IRoutingPolicy::UP
