+++ vespaclient-container-plugin/src/main/java/com/yahoo/document/restapi/resource/DocumentV1ApiHandler.java
@@ -377,10 +377,6 @@ public class DocumentV1ApiHandler extends AbstractRequestHandler {
                                                           .map(parameters::withRoute)
                                                           .orElse(parameters);
                 break;
-            case CONDITION:
-                parameters = getProperty(request, ROUTE).map(parameters::withRoute)
-                                                        .orElse(parameters);
-                break;
             case FIELD_SET:
                 parameters = getProperty(request, FIELD_SET).map(parameters::withFieldSet)
                                                             .orElse(parameters);
@@ -702,7 +698,7 @@ public class DocumentV1ApiHandler extends AbstractRequestHandler {
             this.reader = reader;
         }
 
-        /** Write is complete when we have stored the buffer — call completion handler. */
+        /** Write is complete when we have stored the buffer — call completion handler. */
         @Override
         public void write(ByteBuffer buf, CompletionHandler handler) {
             try {
@@ -948,18 +944,15 @@ public class DocumentV1ApiHandler extends AbstractRequestHandler {
     static class StorageCluster {
 
         private final String name;
-        private final String configId;
         private final Map<String, String> documentBuckets;
 
-        StorageCluster(String name, String configId, Map<String, String> documentBuckets) {
+        StorageCluster(String name, Map<String, String> documentBuckets) {
             this.name = requireNonNull(name);
-            this.configId = requireNonNull(configId);
             this.documentBuckets = Map.copyOf(documentBuckets);
         }
 
         String name() { return name; }
-        String configId() { return configId; }
-        String route() { return "[Storage:cluster=" + name() + ";clusterconfigid=" + configId() + "]"; }
+        String route() { return name() + "-direct"; }
         Optional<String> bucketOf(String documentType) { return Optional.ofNullable(documentBuckets.get(documentType)); }
 
     }
@@ -968,7 +961,6 @@ public class DocumentV1ApiHandler extends AbstractRequestHandler {
         return clusters.storage().stream()
                        .collect(toUnmodifiableMap(storage -> storage.name(),
                                                   storage -> new StorageCluster(storage.name(),
-                                                                                storage.configid(),
                                                                                 buckets.cluster(storage.name())
                                                                                        .documentType().entrySet().stream()
                                                                                        .collect(toMap(entry -> entry.getKey(),
