+++ vespaclient-container-plugin/src/test/java/com/yahoo/document/restapi/resource/DocumentV1ApiTest.java
@@ -113,7 +113,6 @@ public class DocumentV1ApiTest {
     }
 
     final Map<String, StorageCluster> clusters = Map.of("content", new StorageCluster("content",
-                                                                                      "config-id",
                                                                                       Map.of("music", "default")));
     ManualClock clock;
     MockDocumentAccess access;
@@ -139,7 +138,7 @@ public class DocumentV1ApiTest {
     public void testResolveCluster() {
         assertEquals("content",
                      DocumentV1ApiHandler.resolveCluster(Optional.empty(), clusters).name());
-        assertEquals("[Storage:cluster=content;clusterconfigid=config-id]",
+        assertEquals("content-direct",
                      DocumentV1ApiHandler.resolveCluster(Optional.of("content"), clusters).route());
         try {
             DocumentV1ApiHandler.resolveCluster(Optional.empty(), Map.of());
@@ -157,8 +156,8 @@ public class DocumentV1ApiTest {
         }
         try {
             Map<String, StorageCluster> twoClusters = new TreeMap<>();
-            twoClusters.put("one", new StorageCluster("one", "one-config", Map.of()));
-            twoClusters.put("two", new StorageCluster("two", "two-config", Map.of()));
+            twoClusters.put("one", new StorageCluster("one", Map.of()));
+            twoClusters.put("two", new StorageCluster("two", Map.of()));
             DocumentV1ApiHandler.resolveCluster(Optional.empty(), twoClusters);
             fail("More than one cluster and no document type should fail");
         }
@@ -193,7 +192,7 @@ public class DocumentV1ApiTest {
 
         // GET at root is a visit. Numeric parameters have an upper bound.
         access.expect(parameters -> {
-            assertEquals("[Storage:cluster=content;clusterconfigid=config-id]", parameters.getRoute().toString());
+            assertEquals("content-direct", parameters.getRoute().toString());
             assertEquals("default", parameters.getBucketSpace());
             assertEquals(1024, parameters.getMaxTotalHits());
             assertEquals(100, ((StaticThrottlePolicy) parameters.getThrottlePolicy()).getMaxPendingCount());
@@ -275,7 +274,7 @@ public class DocumentV1ApiTest {
         // GET with full document ID is a document get operation which returns 404 when no document is found
         access.session.expect((id, parameters) -> {
             assertEquals(doc1.getId(), id);
-            assertEquals(parameters().withRoute("[Storage:cluster=content;clusterconfigid=config-id]").withFieldSet("go"), parameters);
+            assertEquals(parameters().withRoute("content-direct").withFieldSet("go"), parameters);
             parameters.responseHandler().get().handleResponse(new DocumentResponse(0, null));
             return new Result(Result.ResultType.SUCCESS, null);
         });
