+++ vespaclient-core/src/main/java/com/yahoo/vespaclient/ClusterList.java
@@ -4,40 +4,38 @@ package com.yahoo.vespaclient;
 import com.yahoo.cloud.config.ClusterListConfig;
 import com.yahoo.config.subscription.ConfigGetter;
 
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
 
 /** A list of content clusters, either obtained from a list, a given config or by self-subscribing */
 public class ClusterList {
 
-    List<ClusterDef> contentClusters = new ArrayList<>();
+    private final List<ClusterDef> contentClusters;
 
     public ClusterList() {
-        this(new ArrayList<>());
+        this(List.of());
     }
     
     public ClusterList(List<ClusterDef> contentClusters) {
-        this.contentClusters = contentClusters;
+        this.contentClusters = List.copyOf(contentClusters);
     }
 
     public ClusterList(String configId) {
-        configure(new ConfigGetter<>(ClusterListConfig.class).getConfig(configId));
+        this(new ConfigGetter<>(ClusterListConfig.class).getConfig(configId));
     }
     
     public ClusterList(ClusterListConfig config) {
-        configure(config);
+        this(parse(config));
     }
 
-    private void configure(ClusterListConfig config) {
-        contentClusters.clear(); // TODO: Create a new
-        for (int i = 0; i < config.storage().size(); i++)
-            contentClusters.add(new ClusterDef(config.storage(i).name(), config.storage(i).configid()));
-    }
-
-    /** Returns a reference to the mutable list */
     public List<ClusterDef> getStorageClusters() {
-        return contentClusters; // TODO: Use immutable list
+        return contentClusters;
+    }
+
+    private static List<ClusterDef> parse(ClusterListConfig config) {
+        return config.storage().stream()
+                     .map(storage -> new ClusterDef(storage.name()))
+                     .collect(Collectors.toUnmodifiableList());
     }
 
 }
